Rename dataset files to datatype & rename dataset_base to dataset.
