+++ onmt/datatypes/text_datatype.py
@@ -2,7 +2,7 @@
 
 import torch
 
-from onmt.inputters.dataset_base import DatasetBase
+from onmt.inputters.dataset import DatasetBase
 
 
 class TextDataset(DatasetBase):
